Cah gamemode

+++ Player/Player.py
@@ -3,7 +3,10 @@ Player object, identified by his unique username.
 Stores information about his username, scores etc
 Can later be linked to a database to keep track of (leaderboard/overall scores)
 '''
+import asyncio
 from enum import Enum
+
+import telegram
 
 from Chat.DialogueReader import DialogueReader
 from telegram import InputMediaPhoto
@@ -13,14 +16,45 @@ class Player:
     _chatID = 0
     _score = 0
     _user_data = None
+    _running_edit = False
     
+    def timeOutRetryDecorator(func):
+        async def wrapper(*args, **kwargs):
+            max_retries = 5
+            retry_delay = 1
+            retry_count = 0
+            while retry_count < max_retries:
+                try:
+                    return await func(*args, **kwargs)
+                except telegram.error.TimedOut as e:
+                    if retry_count >= max_retries:
+                        print(f"Timed out: {str(e)} after {retry_count} retries. Giving up.")
+                        raise
+                    else:
+                        print(f"Timed out: {str(e)} running {func.__name__}. Retrying in {retry_delay} seconds...")
+                        await asyncio.sleep(retry_delay)
+                        retry_delay *= 2
+                        retry_count += 1
+
+            return await wrapper(*args, **kwargs)
+        return wrapper
+
     class PlayerConstants(Enum):
+        PRESSING_BUTTON = "pressing_button"
+        WAITING_MSG = "waiting_msg"
         PROMPT = "prompt"
         LIE = "lie"
         NEXT_LIE = "next_lie"
         HAS_VOTED = "has_voted"
+        ARCADE_PROMPT_LIST = "arcade_prompt_list"
+        ARCADE_GEN_STRING = "arcade_gen_string"
+        BANNED_CATEGORY = "banned_category"
+        NEXT_CAPTION = "next_caption"
+        CAPTION = "caption"
+        HAS_PICKED = "has_picked"
+        ARCADE_IMAGE = "arcade_image"
     
-    def __init__(self, username, chatID=0, _user_data={}, score=0, sentPrompt=False):
+    def __init__(self, username, chatID=0, _user_data={}, score=0):
         self._username = username
         self._chatID = chatID
         self._score = score
@@ -63,9 +97,6 @@ class Player:
     
     def getUsername(self):
         return self._username
-
-    def getChatID(self):
-        return self._chatID
     
     def inRoom(self):
         return self._user_data['roomCode'] != ""
@@ -95,12 +126,14 @@ class Player:
             return False
         self._user_data[itemKey.value] = value
 
-    def querySentItem(self, itemKey):
-        if itemKey not in Player.PlayerConstants.__members__.values():
+    def queryItem(self, itemKey):
+        if itemKey not in Player.PlayerConstants.__members__.values() or itemKey.value not in self._user_data:
             return False
         try:
-            return self._user_data[itemKey.value] is not None and self._user_data[itemKey.value] is not False
+            if self._user_data[itemKey.value] is not None and self._user_data[itemKey.value] is not False:
+                return self._user_data[itemKey.value]
         except KeyError:
+            print(f"itemKey {itemKey.value} returned KeyError, likely inputted string instead of player constant")
             return False
     
     async def startGame(self):
@@ -120,38 +153,74 @@ class Player:
             return False
 
     # Methods to send messages
-    async def editMessage(self, messageKey, message, newMessageKey=None, reply_markup=None, parse_mode=None):
-        await self._user_data[messageKey].edit_text(text=DialogueReader.queryDialogue(message), reply_markup=reply_markup, parse_mode=parse_mode)
-        if newMessageKey != None:
-            self._user_data[newMessageKey] = self._user_data.pop(messageKey)
+    # async def editMessage(self, messageKey, message, newMessageKey=None, reply_markup=None, parse_mode=None):
+    #     await self._user_data[messageKey].edit_text(text=DialogueReader.queryDialogue(message), reply_markup=reply_markup, parse_mode=parse_mode)
+    #     if newMessageKey != None:
+    #         self._user_data[newMessageKey] = self._user_data.pop(messageKey)
 
-
+    @timeOutRetryDecorator
     async def editMessage(self, messageKey, message, newMessageKey=None, reply_markup=None, parse_mode=None, **kwargs):
-        await self._user_data[messageKey].edit_text(text=DialogueReader.queryDialogue(message, **kwargs), reply_markup=reply_markup, parse_mode=parse_mode)
+        if messageKey not in self._user_data:
+            return
+        try:
+            await self._user_data[messageKey].edit_text(text=DialogueReader.queryDialogue(message, **kwargs), reply_markup=reply_markup, parse_mode=parse_mode)
+        except telegram.error.BadRequest as badReqError:
+            print(badReqError)
+            pass
         if newMessageKey != None:
             self._user_data[newMessageKey] = self._user_data.pop(messageKey)
 
-    async def editImageURL(self, messageKey, imageURL, newMessageKey=None, reply_markup=None, parse_mode=None):
-        await self._user_data[messageKey].edit_media(media=InputMediaPhoto(imageURL), reply_markup=reply_markup, parse_mode=parse_mode)
+    @timeOutRetryDecorator
+    async def editImageURL(self, messageKey, imageURL, newMessageKey=None, reply_markup=None, parse_mode=None, caption=None, **kwargs):
+        await self._user_data[messageKey].edit_media(media=InputMediaPhoto(imageURL, caption=DialogueReader.queryDialogue(caption, **kwargs), parse_mode=parse_mode), reply_markup=reply_markup)
         if newMessageKey != None:
             self._user_data[newMessageKey] = self._user_data.pop(messageKey)
 
-    async def deleteMessage(self, messageKey):
+    async def deleteMessage(self, messageKey, itemKey=None):
+        if messageKey not in self._user_data:
+            return
+        # Item key is used if the messageKey is a dictionary or list of messages
+        if itemKey != None:
+            await self._user_data[messageKey][itemKey].delete()
+            try:
+                del self._user_data[messageKey][itemKey]
+            except TypeError:
+                pass
+            return
         await self._user_data[messageKey].delete()
         del self._user_data[messageKey]
 
-    # Including a message key will store the message's ID in the user_data which can be editted later
-    async def sendMessage(self, bot, message, messageKey=None, reply_markup=None, raw=False):
-        messasgeID = await DialogueReader.sendMessageByID(bot, self._chatID, message, raw=raw, reply_markup=reply_markup)
-        if messageKey != None:
-            self._user_data[messageKey] = messasgeID
+    @timeOutRetryDecorator
+    async def deleteMessageList(self, messageKeyList):
+        if messageKeyList not in self._user_data:
+            return
+        try:
+            for message in self._user_data[messageKeyList]:
+                try:
+                    await message.delete()
+                except telegram.error.BadRequest as badReqError:
+                    pass
+        except TypeError:
+            pass
+        del self._user_data[messageKeyList]
 
-    async def sendMessage(self, bot, message, messageKey=None, reply_markup=None, raw=False, **kwargs):
-        messasgeID = await DialogueReader.sendMessageByID(bot, self._chatID, message, reply_markup=reply_markup, raw=raw, **kwargs)
+    # # Including a message key will store the message's ID in the user_data which can be editted later
+    # async def sendMessage(self, bot, message, messageKey=None, reply_markup=None, raw=False):
+    #     messasgeID = await DialogueReader.sendMessageByID(bot, self._chatID, message, raw=raw, reply_markup=reply_markup)
+    #     if messageKey != None:
+    #         self._user_data[messageKey] = messasgeID
+
+    async def sendMessage(self, bot, message, messageKey=None, reply_markup=None, raw=False, parse_mode=None, **kwargs):
+        messasgeID = await DialogueReader.sendMessageByID(bot, self._chatID, message, reply_markup=reply_markup, raw=raw, parse_mode=parse_mode, **kwargs)
         if messageKey != None:
             self._user_data[messageKey] = messasgeID
         
-    async def sendImageURL(self, bot, imageURL, messageKey=None, reply_markup=None):
-        messasgeID = await DialogueReader.sendImageURLByID(bot, self._chatID, imageURL, reply_markup=reply_markup)
+    async def sendImageURL(self, bot, imageURL, messageKey=None, reply_markup=None, caption=None, raw=False, parse_mode=None, **kwargs):
+        messasgeID = await DialogueReader.sendImageURLByID(bot, self._chatID, imageURL, caption=caption, raw=raw, parse_mode=parse_mode, reply_markup=reply_markup, **kwargs)
         if messageKey != None:
             self._user_data[messageKey] = messasgeID
+    async def sendMediaGroup(self, bot, mediaGroup, messageKey=None, caption=None, raw=False, parse_mode=None, **kwargs):
+        # Note messageID will be a list of message IDs for each media
+        messasgeIDs = await DialogueReader.sendMediaGroupByID(bot, self._chatID, mediaGroup, caption=caption, raw=raw, parse_mode=parse_mode, **kwargs)
+        if messageKey != None:
+            self._user_data[messageKey] = messasgeIDs
